reorganize default_scenarios, deploy builtin contracts automatically from spam
